+++ src/effect/visual.ts
@@ -1,0 +1,6 @@
+import { Base } from './base'
+
+/**
+ * Modifies the visual contents of a layer.
+ */
+export class Visual extends Base {}